% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_daymonth.R
\name{add_daymonth}
\alias{add_daymonth}
\title{Function to calculate memory used by all objects in the global environment.}
\usage{
add_daymonth(df)
}
\arguments{
\item{type}{defines whether you want the sum of all objects or a sorted data frame}
}
\description{
Function to calculate memory used by all objects in the global environment.
}
