% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_value.R
\name{bin_value}
\alias{bin_value}
\title{Function to calculate memory used by all objects in the global environment.}
\usage{
bin_value(x, bin_res)
}
\arguments{
\item{x}{The variable you want to bin}

\item{bin_res}{The resolution for binning}
}
\description{
Function to calculate memory used by all objects in the global environment.
}
