% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthName.R
\name{binValue}
\alias{binValue}
\title{Bin a value within a vector}
\usage{
binValue(X, Y, method = "lower")
}
\arguments{
\item{x}{A number}
}
\value{
Nearest value to x in vector Y
}
\description{
Takes a number and works out where it lies in another vector
}
