% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extNumFromString.R
\name{doyFromDF}
\alias{doyFromDF}
\title{Create a vector of days of the year from a data frame}
\usage{
doyFromDF(df, yy)
}
\arguments{
\item{df.}{A data frame which has Year, Month and Day in named columns}
}
\value{
A vector with the days of the year in each row
}
\description{
Takes a string and extracts the numbers that appear in it. Useful for cases where you want to pull out years
}
