% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_grid.R
\name{expand_grid}
\alias{expand_grid}
\title{tibble version of expand.grid}
\usage{
expand_grid(..., KEEP.OUT.ATTRS = TRUE, stringsAsFactors = FALSE)
}
\value{
A tibble containing all possible combinations of the given variables
}
\description{
Takes the number of a month and converts it to the name of that month
}
