% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extNumFromString.R
\name{extNumFromString}
\alias{extNumFromString}
\title{Extract the numbers from a string}
\usage{
extNumFromString(x)
}
\arguments{
\item{x}{A string}
}
\value{
the numbers that appear in the string
}
\description{
Takes a string and extracts the numbers that appear in it. Useful for cases where you want to pull out years
}
