% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory.R
\name{global_memory}
\alias{global_memory}
\title{Function to calculate memory used by all objects in the global environment.}
\usage{
global_memory()
}
\arguments{
\item{type}{defines whether you want the sum of all objects or a sorted data frame}
}
\description{
Function to calculate memory used by all objects in the global environment.
}
