% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not_in.R
\name{\%nin\%}
\alias{\%nin\%}
\title{Function to check if a value is in a set}
\usage{
x \%nin\% y
}
\arguments{
\item{x}{vector to check}

\item{y}{set to check if data is in}
}
\description{
Function to check if a value is in a set
}
