% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_2_index.R
\name{int_to_ranking}
\alias{int_to_ranking}
\title{Theme used for blog posts}
\usage{
int_to_ranking(i)
}
\arguments{
\item{i}{Number you want to convert to 1st, 2nd etc.}
}
\description{
Theme used for blogs
}
