% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_files.R
\name{list_files}
\alias{list_files}
\title{Function to calculate memory used by all objects in the global environment.}
\usage{
list_files(x, recursive = TRUE, extension = NULL, full_path = FALSE)
}
\arguments{
\item{x}{Directory you want to look in. This must be stated}

\item{recursive}{Do you want to do a recursive search?}

\item{extension}{filter by specified extension}
}
\value{
list of files
}
\description{
Function to calculate memory used by all objects in the global environment.
}
