% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthName.R
\name{month_name}
\alias{month_name}
\title{Convert number of month to name of month}
\usage{
month_name(mm)
}
\arguments{
\item{mm}{A number of the month}
}
\value{
MM Name of month to return
}
\description{
Takes the number of a month and converts it to the name of that month
}
