% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/month_season.R
\name{month_season}
\alias{month_season}
\title{Convert number of month to season}
\usage{
month_season(mm)
}
\arguments{
\item{mm}{A number of the month}
}
\value{
MM Name of season to return
}
\description{
Takes the number of a month and converts it to the season
}
