% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthName.R
\name{nearestValue}
\alias{nearestValue}
\title{Find the nearest value to x in vector Y}
\usage{
nearestValue(x, Y)
}
\arguments{
\item{x}{A number}
}
\value{
Nearest value to x in vector Y
}
\description{
Takes a number and calculates the nearest value to it in a vector
}
