% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_all.R
\name{read_all}
\alias{read_all}
\title{Read all files in a database}
\usage{
read_all(df, File, read_function)
}
\arguments{
\item{df}{data frame}

\item{File}{Column name for files}

\item{read_function}{name of function to use for reading in}
}
\value{
nested data frame
}
\description{
Takes the number of a month and converts it to the name of that month
}
